stage 4

+++ server/src/main/java/org/apache/iotdb/db/engine/modification/ModificationFile.java
@@ -132,9 +132,8 @@ public class ModificationFile implements AutoCloseable {
 
   public void remove() throws IOException {
     close();
-    File f=FSFactoryProducer.getFSFactory().getFile(filePath);
-    if(f.exists())
-      f.delete();
+    File f = FSFactoryProducer.getFSFactory().getFile(filePath);
+    if (f.exists()) f.delete();
   }
 
   public boolean exists() {
